+++ askbot/doc/source/changelog.rst
@@ -20,6 +20,7 @@ Development version (not released yet)
 * Added test on maximum length of title working for utf-8 text (Evgeny)
 * Added caching and invalidation to the question page (Evgeny)
 * Added a management command delete_contextless_activities (Evgeny)
+* LDAP login configuration (github user `monkut <https://github.com/monkut>`_)
 
 0.7.39 (Jan 11, 2012)
 ---------------------